Freeze the public API. Breaking changes ahead of the freeze (issues #28-#31): Match
columns are all int with a new Matches() accessor; WithSearchConfig merges
field-by-field (no zero-value trap); DefaultToolbarConfig takes a fyne.Window;
the Search vs SearchDebounced contract and SearchMode consts are pinned in godoc.
The exported surface of prettyview + fonttheme is enumerated and frozen by
TestExportedSurfaceGolden, with a Stability note (post-1.0 breaks ship under .../v2).
Also fixes search find-next stuck at match #2 (#34) and aligns the search-bar glyph.
Coverage >90%; govulncheck clean.

+++ wrap_view_test.go
@@ -194,7 +194,7 @@ func TestMatchHighlightOnContinuationRow(t *testing.T) {
 	}
 	mt := pv.search.matches[0]
 	var dst []int32
-	dst = pv.doc.WrapBreaks(mt.Line, dst[:0])
+	dst = pv.doc.WrapBreaks(int32(mt.Line), dst[:0])
 	if len(dst) < 3 {
 		t.Fatal("match line did not wrap")
 	}
@@ -205,7 +205,7 @@ func TestMatchHighlightOnContinuationRow(t *testing.T) {
 	if sub == 0 {
 		t.Fatal("match landed on the first row; fixture too short to test continuation")
 	}
-	wantY := pv.met.RowY(int(pv.doc.FirstVisualRowOfLine(mt.Line)) + sub)
+	wantY := pv.met.RowY(int(pv.doc.FirstVisualRowOfLine(int32(mt.Line))) + sub)
 	found := false
 	for _, rc := range pv.r.matchRects {
 		if rc.Visible() {
